+++ src/main/java/com/yuuko/core/commands/moderation/commands/NukeCommand.java
@@ -25,6 +25,11 @@ public class NukeCommand extends Command {
 
     @Override
     public void onCommand(MessageEvent e) {
+        // Manual rate limiting for nuke - default 5 seconds.
+        if(!checkCooldown(e, 5000L)) {
+            return;
+        }
+
         List<TextChannel> channels = e.getMessage().getMentionedChannels();
         if(channels.size() > 0) {
             TextChannel channel = channels.get(0);
@@ -58,11 +63,15 @@ public class NukeCommand extends Command {
             if(sortedMessages.get(false).size() > 2) {
                 e.getChannel().deleteMessages(sortedMessages.get(false).subList(1, sortedMessages.get(false).size())).queue(s -> {
                     ModerationLogSetting.execute(e, messages.size()); // Attempt to add event to moderation log.
-                }, f -> log.warn("An error occurred while running the {} class, message: {}", this.getClass().getSimpleName(), f.getMessage()));
+                }, f -> {
+                    log.warn("An error occurred while running the {} class, message: {}", this.getClass().getSimpleName(), f.getMessage());
+                });
             }
 
             // Removes messages that are too old to be mass-deleted.
-            sortedMessages.get(true).listIterator().forEachRemaining(message -> message.delete().queue(s -> {}, f -> log.warn("An error occurred while running the {} class, message: {}", this.getClass().getSimpleName(), f.getMessage(), f)));
+            sortedMessages.get(true).listIterator().forEachRemaining(message -> message.delete().queue(s -> {}, f -> {
+                log.warn("An error occurred while running the {} class, message: {}", this.getClass().getSimpleName(), f.getMessage());
+            }));
         });
     }